+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 ### Features
 
 - Improve Mnemonic Phrase Input UX [PR 2280](https://github.com/input-output-hk/daedalus/pull/2280)
+- Added wallet public key viewing feature ([PR 2271](https://github.com/input-output-hk/daedalus/pull/2271))
 
 ### Fixes